requirements.txt
openai
langchain
langchain-community
langchain-openai
langgraph
python-dotenv
pypdf
requests
pandas
semantic_router
pinecone
serpapi
google-search-results
pyautogen